Moving artifact classes from build to artifacts.
