+++ cibyl/sources/zuul/apis/__init__.py
@@ -20,7 +20,7 @@ from deprecation import deprecated
 
 from cibyl.exceptions.source import SourceException
 from cibyl.sources.zuul.apis.providers import JobsProvider, PipelinesProvider
-from cibyl.sources.zuul.apis.utils.builds import Artifact, ArtifactKind
+from cibyl.sources.zuul.apis.utils.artifacts import Artifact, ArtifactKind
 from cibyl.utils.io import Closeable
