+++ tests/cibyl/unit/sources/zuul/apis/utils/artifacts/test_manifest.py
@@ -18,9 +18,9 @@ from unittest.mock import Mock
 
 import cibyl
 from cibyl.sources.zuul.apis import ArtifactKind
+from cibyl.sources.zuul.apis.utils.artifacts import ArtifactError
 from cibyl.sources.zuul.apis.utils.artifacts.manifest import (
     Manifest, ManifestDownloader)
-from cibyl.sources.zuul.apis.utils.builds import ArtifactError
 
 
 class TestManifestDownloader(TestCase):
